--- project/frontend/templates/game-page.html
@@ -1,15 +1,0 @@
-<!-- game-page -->
-<div id="game-page" style="display: none;">
-<!-- <div style="background-color: bisque;"> -->
-	<center>
-			<canvas id="gameArea" width="1024" height="600"></canvas>
-			<script src="./js/gameUI/utils.js"></script>
-			<script src="./js/gameUI/main-menu.js"></script>
-			<script src="./js/gameUI/tournament-board.js"></script>
-			<script src="./js/gameUI/players-board.js"></script>
-			<script src="./js/gameUI/online-menu.js"></script>
-			<script src="./js/gameUI/online-board.js"></script>
-			<!-- <script src="./js/gamePlay/Main.js" type="module"></script> -->
-	</center>
-<!-- </div> -->
-</div>
